revert_me: Temporarily disabling deployment superpanics
Gcp, aws, cf and specifically npm are all having outages. This causes
superpanic to trigger. We know deployments are broken, since the
infrastructure that the backend is built on is broken and I'd rather not
be woken up all night just waiting for some engineer with 5 times my
salary to fix it.