Improve app-version wait loop
The wait loop that waits for apps to be published and at the current
version does did not handle the case of the app not being propagated to
Edge servers yet , in which case the version header would be missing
entirely, leading to an immediate abort.

This change now also waits for Edge to recognise the app.